Move stuff from __init__.py to scraping_class.py
